refactor: 중복 검사 로직을 validator로 분리해 서비스 계층 책임을 줄임

+++ src/main/java/com/peeerr/climbing/service/CommentService.java
@@ -44,14 +44,14 @@ public class CommentService {
 
     public void editComment(Long commentId, CommentEditRequest request, Long loginId) {
         Comment comment = getCommentById(commentId);
-        comment.getMember().checkOwner(loginId);
+        comment.checkOwner(loginId);
 
         comment.changeContent(request.getContent());
     }
 
     public void removeComment(Long commentId, Long loginId) {
         Comment comment = getCommentById(commentId);
-        comment.getMember().checkOwner(loginId);
+        comment.checkOwner(loginId);
 
         commentRepository.delete(comment);
     }
